ref: Move interactive viewer inside axis chart scaffold to make it more reusable
