+++ lib/src/chart/scatter_chart/scatter_chart.dart
@@ -53,7 +53,7 @@ class _ScatterChartState extends AnimatedWidgetBaseState<ScatterChart> {
 
     return AxisChartScaffoldWidget(
       data: showingData,
-      chart: ScatterChartLeaf(
+      chartBuilder: (context, chartRect) => ScatterChartLeaf(
         data:
             _withTouchedIndicators(_scatterChartDataTween!.evaluate(animation)),
         targetData: _withTouchedIndicators(showingData),
